requirements.txt
ollama
pdfplumber
pytesseract
Pillow
